minor changes to satisfy issues flagged by Coverity Scan

+++ src/sflow/sflow_poller.c
@@ -68,7 +68,7 @@ void sfl_poller_set_sFlowCpReceiver(SFLPoller *poller, uint32_t sFlowCpReceiver)
 }
 
 uint32_t sfl_poller_get_sFlowCpInterval(SFLPoller *poller) {
-  return (uint32_t)poller->sFlowCpInterval;
+  return poller->sFlowCpInterval;
 }
 
 void sfl_poller_set_sFlowCpInterval(SFLPoller *poller, uint32_t sFlowCpInterval) {
